+++ test/unit/gen_class/Mail_add_2_msg.py
@@ -60,6 +60,7 @@ class UnitTest(unittest.TestCase):
         """
 
         self.to = ["mail_address@domain.name"]
+        self.msg = "Test email line"
 
     def test_non_string2(self):
 
@@ -71,16 +72,13 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = "Test email line"
         msg2 = {1: "Test ", 2: "email ", 3: "line2"}
-
         email = gen_class.Mail(self.to)
-
-        email.add_2_msg(msg)
+        email.add_2_msg(self.msg)
         email.add_2_msg(msg2)
 
         self.assertEqual((email.to, email.msg), (
-            self.to, msg + '{"1": "Test ", "2": "email ", "3": "line2"}'))
+            self.to, self.msg + '{"1": "Test ", "2": "email ", "3": "line2"}'))
 
     def test_non_string(self):
 
@@ -92,16 +90,13 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = "Test email line"
         msg2 = ["Test ", "email ", "line2"]
-
         email = gen_class.Mail(self.to)
-
-        email.add_2_msg(msg)
+        email.add_2_msg(self.msg)
         email.add_2_msg(msg2)
 
         self.assertEqual((email.to, email.msg),
-                         (self.to, msg + '["Test ", "email ", "line2"]'))
+                         (self.to, self.msg + '["Test ", "email ", "line2"]'))
 
     def test_empty_test(self):
 
@@ -113,14 +108,11 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = "Test email line"
-
         email = gen_class.Mail(self.to)
-
-        email.add_2_msg(msg)
+        email.add_2_msg(self.msg)
         email.add_2_msg()
 
-        self.assertEqual((email.to, email.msg), (self.to, msg))
+        self.assertEqual((email.to, email.msg), (self.to, self.msg))
 
     def test_add_exist(self):
 
@@ -132,15 +124,12 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = "Test email line"
         msg2 = "Test email line2"
-
         email = gen_class.Mail(self.to)
-
-        email.add_2_msg(msg)
+        email.add_2_msg(self.msg)
         email.add_2_msg(msg2)
 
-        self.assertEqual((email.to, email.msg), (self.to, msg + msg2))
+        self.assertEqual((email.to, email.msg), (self.to, self.msg + msg2))
 
     def test_initial_add(self):
 
@@ -152,13 +141,10 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        msg = "Test email line"
-
         email = gen_class.Mail(self.to)
+        email.add_2_msg(self.msg)
 
-        email.add_2_msg(msg)
-
-        self.assertEqual((email.to, email.msg), (self.to, msg))
+        self.assertEqual((email.to, email.msg), (self.to, self.msg))
 
 
 if __name__ == "__main__":